Add ability to specify binary via command line args
Adds the ability to specify a firmware binary via command line arguments with argparse.
This is done by running the application with -fw or --firmware followed by the path to the binary.
This allows the application to be used in a manner where the binary file can be "preloaded".

In the event of unrecognized arguments, or a nonexistent file, an appropriate error message is shown.

Additionally, the README.md has been updated to reflect this change.